+++ support/example/json/codeGo/item.go
@@ -5,7 +5,7 @@ package sheeterJson
 
 type Item struct {
 	// 物品數量
-	Count int64 `json:"Count"`
+	Count int32 `json:"Count"`
 	// 物品編號
-	ItemID int64 `json:"ItemID"`
+	ItemID int32 `json:"ItemID"`
 }
